Plasmid dilution series results

These are the results data from the pds_raw data as calculated by the BioRad QX100 Droplet Digital PCR System.

Format

A data frame with 64 observations on the following 44 variables.

Well

a factor with levels A01 to H04

ExptType

a factor with levels Absolute Quantification

Experiment

a factor with levels ABS

Sample

a factor with levels B B + P 10^2 gDNA gDNA + P 10^0 gDNA + P 10^1 gDNA + P 10^-1 gDNA + P 10^2 gDNA + P 10^3 gDNA + P 10^4

TypeAssay

a factor with levels Ch1NTC Ch1Unknown Ch2NTC Ch2Unknown

Assay

a factor with levels ileS styA

Status

a factor with levels Manual

Concentration

a numeric vector

TotalConfMax

a logical vector

TotalConfMin

a logical vector

PoissonConfMax

a numeric vector

PoissonConfMin

a numeric vector

Positives

a numeric vector

Negatives

a numeric vector

Ch1.Ch2.

a numeric vector

Ch1.Ch2..1

a numeric vector

Ch1.Ch2..2

a numeric vector

Ch1.Ch2..3

a numeric vector

Linkage

a numeric vector

AcceptedDroplets

a numeric vector

CNV

a logical vector

TotalCNVMax

a logical vector

TotalCNVMin

a logical vector

PoissonCNVMax

a logical vector

PoissonCNVMin

a logical vector

ReferenceCopies

a logical vector

UnknownCopies

a logical vector

Ratio

a numeric vector

TotalRatioMax

a logical vector

TotalRatioMin

a logical vector

PoissonRatioMax

a numeric vector

PoissonRatioMin

a numeric vector

FractionalAbundance

a numeric vector

TotalFractionalAbundanceMax

a logical vector

TotalFractionalAbundanceMin

a logical vector

PoissonFractionalAbundanceMax

a numeric vector

PoissonFractionalAbundanceMin

a numeric vector

ReferenceAssayNumber

a numeric vector

TargetAssayNumber

a numeric vector

MeanAmplitudeofPositives

a numeric vector

MeanAmplitudeofNegatives

a numeric vector

MeanAmplitudeTotal

a numeric vector

ExperimentComments

a logical vector

MergedWells

a logical vector

Source

Michael Jahn Flow cytometry group / Environmental microbiology Helmholtz Centre for Environmental Research - UFZ Permoserstrasse 15 / 04318 Leipzig / Germany phone +49 341 235 1318 michael.jahn [at] ufz.de / www.ufz.de

Details

Setup: Duplex assay with constant amount of genomic DNA and six 10-fold dilutions of plasmid DNA with 4 replicates, ranging theoretically from ~ 10^4 to 10^-1 copies/ micro L plus 4 replicates without plasmid DNA. Included are No-gDNA-control and No-template-control, 2 replicates each.

Annotation: FX.Y (X = dilution number, Y = replicate number). Hardware: Bio-Rad QX100 Droplet digital PCR system Details: Genomic DNA isolated from Pseudomonas putida KT2440. Plasmid is pCOM10-StyA::EGFP StyB [Jahn et al., 2013, Curr Opin Biotechnol, Vol. 24 (1): 79-87]. Template DNA was heat treated at 95 degree Celsius for 5 min prior to PCR. Channel 1, primers for genomic DNA marker ileS, Taqman probes (FAM labelled). Channel 2, primers for plasmid DNA marker styA, Taqman probes (HEX labelled).

References

Jahn et al., 2013, Curr Opin Biotechnol, Vol. 24 (1): 79-87

Examples

summary(extract_run(read_QX100(pds), 1L:10))

#> The assumed volume of partitions in each run is equal to 0.85.

#> The assumed volume uncertainty in each run is equal to 0.017.

#> 
#> Number of positive partitions: 401, 651, 422, 133, ... 
#> Total number of partitions:    11964, 11592, 10903, 11196, ... 
#> 
#> Number of runs:        10 
#> Number of experiments: 1 
#> 
#>  experiment         replicate assay method     lambda  lambda.low  lambda.up
#>  Ch1Unknown A01.gDNA + P 10^4  ileS   dube 0.03409180 0.030760423 0.03743430
#>  Ch1Unknown A01.gDNA + P 10^4  ileS   bhat 0.03409180 0.032389250 0.03579434
#>  Ch1Unknown B01.gDNA + P 10^4  ileS   dube 0.05779800 0.053367344 0.06224838
#>  Ch1Unknown B01.gDNA + P 10^4  ileS   bhat 0.05779800 0.055532407 0.06006360
#>  Ch1Unknown C01.gDNA + P 10^4  ileS   dube 0.03947389 0.035714531 0.04324743
#>  Ch1Unknown C01.gDNA + P 10^4  ileS   bhat 0.03947389 0.037552203 0.04139557
#>  Ch1Unknown D01.gDNA + P 10^4  ileS   dube 0.01195036 0.009921443 0.01398341
#>  Ch1Unknown D01.gDNA + P 10^4  ileS   bhat 0.01195036 0.010914130 0.01298660
#>  Ch1Unknown E01.gDNA + P 10^3  ileS   dube 0.09467602 0.089011882 0.10037243
#>  Ch1Unknown E01.gDNA + P 10^3  ileS   bhat 0.09467602 0.091777903 0.09757415
#>  Ch1Unknown F01.gDNA + P 10^3  ileS   dube 0.08554124 0.080246554 0.09086412
#>  Ch1Unknown F01.gDNA + P 10^3  ileS   bhat 0.08554124 0.082832658 0.08824983
#>  Ch1Unknown G01.gDNA + P 10^3  ileS   dube 0.09421145 0.088126702 0.10033345
#>  Ch1Unknown G01.gDNA + P 10^3  ileS   bhat 0.09421145 0.091097466 0.09732544
#>  Ch1Unknown H01.gDNA + P 10^3  ileS   dube 0.07943170 0.074357306 0.08453197
#>  Ch1Unknown H01.gDNA + P 10^3  ileS   bhat 0.07943170 0.076836095 0.08202730
#>  Ch1Unknown A02.gDNA + P 10^2  ileS   dube 0.07701105 0.071784275 0.08226529
#>  Ch1Unknown A02.gDNA + P 10^2  ileS   bhat 0.07701105 0.074337298 0.07968480
#>  Ch1Unknown B02.gDNA + P 10^2  ileS   dube 0.09606164 0.090328139 0.10182820
#>  Ch1Unknown B02.gDNA + P 10^2  ileS   bhat 0.09606164 0.093127926 0.09899535
#>          m     m.low      m.up          c      c.low       c.up    k     n
#>   407.8742  368.0177  447.8640 0.04010799 0.03618873 0.04404036  401 11964
#>   407.8742  387.5050  428.2435 0.04010799 0.03795035 0.04226564  401 11964
#>   669.9945  618.6343  721.5833 0.06799765 0.06278511 0.07323339  651 11592
#>   669.9945  643.7317  696.2573 0.06799765 0.06500535 0.07098996  651 11592
#>   430.3838  389.3955  471.5267 0.04643987 0.04201710 0.05087933  422 10903
#>   430.3838  409.4317  451.3359 0.04643987 0.04399571 0.04888402  422 10903
#>   133.7963  111.0805  156.5583 0.01405925 0.01167229 0.01645107  133 11196
#>   133.7963  122.1946  145.3980 0.01405925 0.01280815 0.01531036  133 11196
#>  1119.3546 1052.3875 1186.7033 0.11138356 0.10471986 0.11808521 1068 11823
#>  1119.3546 1085.0901 1153.6191 0.11138356 0.10731077 0.11545634 1068 11823
#>  1041.2936  976.8413 1106.0889 0.10063676 0.09440771 0.10689896  998 12173
#>  1041.2936 1008.3219 1074.2652 0.10063676 0.09686776 0.10440576  998 12173
#>   959.8263  897.8348 1022.1972 0.11083700 0.10367847 0.11803935  916 10188
#>   959.8263  928.1010  991.5515 0.11083700 0.10655503 0.11511897  916 10188
#>   974.7064  912.4385 1037.2918 0.09344906 0.08747918 0.09944938  937 12271
#>   974.7064  942.8557 1006.5570 0.09344906 0.08986885 0.09702926  937 12271
#>   862.3698  803.8403  921.2067 0.09060124 0.08445209 0.09678269  830 11198
#>   862.3698  832.4291  892.3104 0.09060124 0.08697106 0.09423141  830 11198
#>  1125.3621 1058.1941 1192.9173 0.11301369 0.10626840 0.11979788 1073 11715
#>  1125.3621 1090.9937 1159.7305 0.11301369 0.10888802 0.11713936 1073 11715